Made Selector. Can add to cart and update quantity in bevdetail.
